Refactor mobile header component: move to new directory, update styles, and replace Instagram icon
