+++ app/component/HeaderMobile.module.css
@@ -1,0 +1,4 @@
+.headerMobile{
+    background-color: #FCEFEF;
+  }
+
