+++ app/globals.css
@@ -15,6 +15,7 @@
 }
 
 body {
+  margin: 0;
   color: var(--foreground);
   background: var(--background);
   font-family: "Titillium Web", Arial, Helvetica, sans-serif;
